feat: implements validation for date type in deep
Before it was checking only if the received date in /batida endpoint
is valid, but it needs to check if is in correct format.
Before, if someone passes a valid date like "2023-12-01", it would be
accepted and it would save the time register as null in database.
This would cause an error when compare data.