fix(report): move owed and exceeded times calculation to outside of total worked hours
The owed and exceeded hours were calculating togheter with
total worked hours, but this was wrong because if x day the person
exceeded time and then x + 1 day the person owed a time, those
values must cancel each other, before the changes they are sum
for every register found. This commit fix this.